+++ GameDesigner/Example~/ExampleServer~/Example2/DB/Example2DBEvent.cs
@@ -21,10 +21,10 @@ namespace Example2
         /// 当服务器属性同步给客户端, 如果需要同步属性到客户端, 需要监听此事件, 并且调用发送给客户端
         /// 参数1: 要发送给哪个客户端
         /// 参数2: cmd
-        /// 参数3: methodHash
+        /// 参数3: protocol
         /// 参数4: pars
         /// </summary>
-        public static System.Action<Net.Server.NetPlayer, byte, ushort, object[]> OnSyncProperty;
+        public static System.Action<Net.Server.NetPlayer, byte, uint, object[]> OnSyncProperty;
 
         /// <summary>
         /// 当实体行属性更改时触发
